Fix:
- ファイル分割ではなくcreate_response.cpp内に関数をまとめた（YAGNI 原則）
- rebaseした影響でコンパイルやができないやgit addした時に.historyが追加されるので以下を修正。
	- class httpのメンバー関数とmapの追加
	- .gitignoreに.historyを追加
